v4l: xilinx: dma: Add interlaced support
This patch adds interlaced support to the Xilinx V4L dma client. In case
of capture pipeline, the field id is read from the callback.

A check is present to find and correct sequence number in case
a frame is dropped i.e. fid is repeated. For this the prev_fid member is
used to store the previous fid value to be compared with one returned.

In case of output pipeline, the field id is set per dma descriptor.

A dma descriptor pointer is added to xvip_dma_buffer so that it may be
passed as a reference while getting the fid.

The video node gets the field type of the subdev prior and checks if it
is of V4L2_FIELD_ALTERNATE type. If yes then height is halved.

Some other minor fixes for checkpatch are also applied here.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
[michals:
v6.12 update
xilinx: v4l: dma: Fix bug in calculating interlaced height
]