dt-bindings: usb: add quirk for xhci reset on resume
Some controllers might lose power during suspend, so they are requires
re-initialization. During the suspend cycle when USB is not pluh-in or
USB not set as a wakeup source then power management turn off the clock
and power, to save power.

To address above suspend-resume use-case snps,xhci-reset-on-resume in
snps,dwc3.yaml and xhci-reset-on-resume quirk in usb-xhci.yaml are
added, which will reset the host controller on resume.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending