spi: spi-zynqmp-gqspi: Send opcode on single line
Depending on the tx/rx bus width the framework sends 1-1-1, 1-1-2, 1-2-2,
1-1-4, or 1-4-4 mode commands. In all the supported modes, irrespective of
tx bus width, the opcode should always be sent on single line, the address
and data need to be send on single/dual/quad lines as per tx bus width
settings. But when tx bus width is dual or quad, then driver was sending
the opcode on dual or quad lines, due to this read write failures were
observed. Update the driver to send opcode on single line, irrespective
of tx bus width.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending