spi: spi-zynqmp-gqspi: Fix erase failure when tx buswidth is 4
Irrespective of the tx buswidth, for erase commands the flash device
expects to receive the address on single line. But when tx buswidth
is set to 4 and the driver issues erase command, it sends the address
on 4 lines. This causes the erase operation to fail.
Update the driver to send the address, for all erase commands, on
single line.

Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending