pinctrl: pinctrl-zynqmp: Add Versal solution without PM support
This is a temporary solution which will be removed in 2026.1 release.
PM support to get the MIO IDs is added in 2024.2 release. So, to give
enough time to migrate to the latest PLM and not to break the existing
customer solutions, maintain the legacy/temporary solution (hardcoded
ID information in the driver) in the driver till 2025.2 release and this
will be removed in 2026.1 release.

Signed-off-by: Sai Krishna Potthuri <sai.krishna.potthuri@amd.com>
Reason: temporary solution, will be removed in 2026.1
State: not-upstreamable
Message-ID: <20241206090010.2328786-1-sai.krishna.potthuri@amd.com>